Change Dump, Mainly Static Code Analysis Changes
Altered a bunch of misdefined and/or misdesigned code pointed out by
static code analysis. These fixes mainly improve game stability,
especially fixes related to memory allocation, but some increase
performance by pruning out useless code.

Fixed the fact that the game will predict a stalemate when two melee
units would both die from combat, instead of predicting a victory for
the unit with the least amount of total damage after combat.

Changed the location of the heavy charge strength buff (units with heavy
charge attacking a unit with all of their potential retreat tiles
blocked get a 1.5x strength boost) so that it is properly included in
combat previewing algorithms.

Recapturing a city no longer causes population or building loss, as was
originally intended by Firaxis.

The free experience function has been fixed so that wonders that award
free experience globally (i.e. not just in the city they are built in)
work as intended.

Tech costs in teamer games will scale with the total amount of cities in
the team instead of individual city counts for each player. Player-based
cost reductions are applied to the whole team based on what fraction of
the team's total cities are owned by the player(s) with reductions.

Fixed the bug where the tech cost discount from other (met) teams
already knowing a tech did not max out at the intended maximum of 30%,
but instead was restricted to 15% in 2-team games, 20% in 3-team games,
22.5% in 4-team games, 24% in 5-team games, 25% in 6-team games, etc. To
keep discounts the same as they were before in the game standard (6
player FFA), the max discount has been lowered from 30% to 25%.

The tech cost discount awarded based on how many (met) teams already
have that tech can now be toggled via an in-game option.

Restored the malus to coup chance if a spy from the CS's ally is present
in the city. This was most likely originally disabled by mistake.

+++ CvGameCoreDLL_Expansion2/CvRandom.cpp
@@ -204,9 +204,9 @@ unsigned short CvRandom::get(unsigned short usNum, const char* pszLog)
 				{
 					char szOut[1024] = {0};
 #ifdef AUI_USE_SFMT_RNG
-					sprintf_s(szOut, "%d, %d, %u, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, uiRtnValue, m_ulRandomSeed, m_ulCallCount, (uint)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
+					sprintf_s(szOut, "%d, %d, %u, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, uiRtnValue, m_ulRandomSeed, m_ulCallCount, (intptr_t)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
 #elif defined(AUI_WARNING_FIXES)
-					sprintf_s(szOut, "%d, %d, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, (uint)us, getSeed(), (uint)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
+					sprintf_s(szOut, "%d, %d, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, (uint)us, getSeed(), (intptr_t)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
 #else
 					sprintf_s(szOut, "%d, %d, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), (uint)usNum, (uint)us, getSeed(), (uint)this, m_bSynchronous?"sync":"async", (pszLog != NULL)?pszLog:"Unknown");
 #endif
@@ -306,9 +306,9 @@ unsigned int CvRandom::getBinom(unsigned int uiNum, const char* pszLog)
 				{
 					char szOut[1024] = { 0 };
 #ifdef AUI_USE_SFMT_RNG
-					sprintf_s(szOut, "%d, %d, %u, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, uiRtnValue, m_ulRandomSeed, m_ulCallCount, (uint)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
+					sprintf_s(szOut, "%d, %d, %u, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, uiRtnValue, m_ulRandomSeed, m_ulCallCount, (intptr_t)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
 #else
-					sprintf_s(szOut, "%d, %d, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, (uint)usRet, getSeed(), (uint)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
+					sprintf_s(szOut, "%d, %d, %u, %u, %u, %8x, %s, %s\n", kGame.getGameTurn(), kGame.getTurnSlice(), uiNum, (uint)usRet, getSeed(), (intptr_t)this, m_bSynchronous ? "sync" : "async", (pszLog != NULL) ? pszLog : "Unknown");
 #endif
 					pLog->Msg(szOut);
